requirements.txt
matplotlib==3.4.2
pytest-cov==2.12.1
tensorflow==2.5
tensorflow_datasets==4.3.0
tensorflow_text==2.5
